m5t1: Дорабатывает логику проигрывания превью фильма в карточке фильма
