+++ project/src/components/films-list/films-list-component.tsx
@@ -1,4 +1,3 @@
-import { useState } from 'react';
 import { Film } from '../../types/film';
 import FilmCardComponent from '../film-card/film-card-component';
 
@@ -6,33 +5,8 @@ type FilmListComponentProps = {
   films: Film[];
 }
 
-type cardState = {
-  isPlaying: boolean;
-  currentId: string | number | null;
-}
-
 // #TODO Необходимо реализовать управление количеством отрисовываемых карточек
 function FilmListComponent({films}: FilmListComponentProps): JSX.Element {
-  const [currentCard, setCurrentCard] = useState<cardState>({
-    currentId: null,
-    isPlaying: false
-  });
-  let hoverTimerID: NodeJS.Timeout | undefined;
-
-  const handleMouseOver = (id: string | number) => {
-    // setCurrentCard((prevState) => ({ ...prevState, currentId: id }));
-    hoverTimerID = setTimeout(() => {
-      setCurrentCard(() => ({ currentId: id, isPlaying: true }));
-    }, 2000);
-  };
-
-  const handleMouseOut = () => {
-    if (currentCard.isPlaying === false) {
-      clearTimeout(hoverTimerID);
-    }
-    setCurrentCard(() => ({ currentId: null, isPlaying: false }));
-  };
-
   return (
     <div className="catalog__films-list">
       {films.map((film) => {
@@ -44,10 +18,6 @@ function FilmListComponent({films}: FilmListComponentProps): JSX.Element {
             name={name}
             videoSrc={videoSrc}
             posterSrc={posterSrc}
-            isPlaying={currentCard.isPlaying}
-            currentId={currentCard.currentId}
-            onMouseOver={handleMouseOver}
-            onMouseOut={handleMouseOut}
           />
         );
       })}
